move Setting function into its own package
